+++ Cantera/src/thermo/ShomateThermo.h
@@ -159,7 +159,7 @@ namespace Cantera {
      * @see ShomatePoly
      * @see ShomatePoly2
      */
-    virtual void install(std::string name, size_t index, size_t type, 
+    virtual void install(std::string name, size_t index, int type,
 			 const doublereal* c,
 			 doublereal minTemp, doublereal maxTemp, 
 			 doublereal refPressure) {
